# package/openapi.yaml
openapi: 3.1.0
info:
  title: My Tasco Secure Agentic RAG API
  version: 1.0.0
  description: |
    Full public contract for the current secure RAG prototype and planned Agentic RAG platform.
    Every operation declares x-implementation-status. Planned operations are contracts, not claims
    that the current backend implements them. OpenAI is an internal outbound dependency and is
    never proxied directly to clients.
  contact:
    name: My Tasco AI Workspace Team
servers:
  - url: http://127.0.0.1:8000
    description: Local prototype
  - url: https://hackathon.example.com/mytasco/v1
    description: Configurable deployment facade
...
paths:
...
  /mytasco/v1/aiwsp/actions/{actionId}:
    get:
      tags: [Actions]
      operationId: getActionPreview
      summary: Get immutable action preview
      x-implementation-status: implemented-core
      parameters:
        - {name: actionId, in: path, required: true, schema: {type: string, format: uuid}}
      responses:
        '200': {description: Action preview, content: {application/json: {schema: {$ref: '#/components/schemas/ActionEnvelope'}}}}
        '404': {$ref: '#/components/responses/NotFound'}
  /mytasco/v1/aiwsp/actions/{actionId}/confirm:
    post:
      tags: [Actions]
      operationId: confirmAction
      summary: Confirm and execute an immutable action
      x-implementation-status: implemented-core
      parameters:
        - $ref: '#/components/parameters/RequestId'
        - $ref: '#/components/parameters/IdempotencyKey'
        - {name: actionId, in: path, required: true, schema: {type: string, format: uuid}}
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [confirmationToken]
              properties:
                confirmationToken: {type: string, minLength: 20}
                stepUpToken: {type: string, writeOnly: true}
      responses:
        '200': {description: Idempotent execution result, content: {application/json: {schema: {$ref: '#/components/schemas/ActionEnvelope'}}}}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '409': {$ref: '#/components/responses/Conflict'}
        '410': {description: Confirmation expired, content: {application/json: {schema: {$ref: '#/components/schemas/ErrorEnvelope'}}}}
  /mytasco/v1/aiwsp/actions/{actionId}/reject:
    post:
      tags: [Actions]
      operationId: rejectAction
      summary: Reject an action preview without upstream mutation
      x-implementation-status: implemented-core
      parameters:
        - {name: actionId, in: path, required: true, schema: {type: string, format: uuid}}
      responses:
        '200': {description: Action rejected, content: {application/json: {schema: {$ref: '#/components/schemas/ActionEnvelope'}}}}
        '404': {$ref: '#/components/responses/NotFound'}
...
  /mytasco/v1/aiwsp/permissions/explain:
    get:
      tags: [Governance, Legacy]
      operationId: explainPermission
      summary: Explain an allow/deny decision without leaking denied metadata
      x-implementation-status: implemented-core
      parameters:
        - {name: documentId, in: query, required: true, schema: {type: string}}
        - $ref: '#/components/parameters/RequestId'
      responses:
        '200': {description: Safe permission explanation, content: {application/json: {schema: {$ref: '#/components/schemas/GenericEnvelope'}}}}
  /mytasco/v1/aiwsp/audit/recent:
    get:
      tags: [Governance, Legacy]
      operationId: getRecentAudit
      summary: Get recent content-free audit events
      x-implementation-status: implemented-core
      responses:
        '200': {description: Audit events, content: {application/json: {schema: {$ref: '#/components/schemas/GenericEnvelope'}}}}
        '403': {$ref: '#/components/responses/Forbidden'}
  /mytasco/v1/aiwsp/admin/traces/{traceId}:
    get:
      tags: [Governance]
      operationId: getTrace
      summary: Get authorized high-level trace without chain-of-thought
      x-implementation-status: implemented-core
      parameters:
        - {name: traceId, in: path, required: true, schema: {type: string, format: uuid}}
      responses:
        '200': {description: Trace metadata, content: {application/json: {schema: {$ref: '#/components/schemas/GenericEnvelope'}}}}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
  /mytasco/v1/aiwsp/admin/security-events:
    get:
      tags: [Governance]
      operationId: listSecurityEvents
      summary: List authorized security events
      x-implementation-status: implemented-core
      responses:
        '200': {description: Security events, content: {application/json: {schema: {$ref: '#/components/schemas/GenericEnvelope'}}}}
        '403': {$ref: '#/components/responses/Forbidden'}
  /mytasco/v1/aiwsp/evaluation/public:
    post:
      tags: [Evaluation, Legacy]
      operationId: runPublicEvaluation
      summary: Run the provided public evaluation set
      x-implementation-status: implemented-core
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                limit: {type: integer, minimum: 1, maximum: 50}
      responses:
        '200': {description: Evaluation results, content: {application/json: {schema: {$ref: '#/components/schemas/GenericEnvelope'}}}}
  /mytasco/v1/aiwsp/evaluations/runs:
    post:
      tags: [Evaluation]
      operationId: createEvaluationRun
      summary: Start a versioned quality/security evaluation
      x-implementation-status: implemented-core
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [datasetId]
              properties:
                datasetId: {type: string}
                candidateConfigId: {type: string}
                baselineConfigId: {type: string}
      responses:
        '202': {description: Evaluation accepted, content: {application/json: {schema: {$ref: '#/components/schemas/GenericEnvelope'}}}}
...
components:
  securitySchemes:
    BearerAuth: {type: http, scheme: bearer, bearerFormat: JWT}
    AppCode: {type: apiKey, in: header, name: X-App-Code, description: Must equal MYTASCO}
  parameters:
    RequestId:
      name: X-Request-Id
      in: header
      required: false
      schema: {type: string, format: uuid}
...
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema: {type: string, minLength: 16, maxLength: 128}
  responses:
...
    Unauthorized: {description: Missing or invalid authentication, content: {application/json: {schema: {$ref: '#/components/schemas/ErrorEnvelope'}}}}
    Forbidden: {description: Authenticated but not authorized, content: {application/json: {schema: {$ref: '#/components/schemas/ErrorEnvelope'}}}}
    NotFound: {description: Resource not found or safely concealed, content: {application/json: {schema: {$ref: '#/components/schemas/ErrorEnvelope'}}}}
    Conflict: {description: Idempotency, confirmation, or resource-version conflict, content: {application/json: {schema: {$ref: '#/components/schemas/ErrorEnvelope'}}}}
...
  schemas:
...
    ErrorEnvelope:
      type: object
      required: [status, code, message, requestId]
      properties:
        status: {type: string, const: error}
        code: {type: string, enum: [invalid_request, unauthorized, forbidden, not_found, conflict, expired, timeout, rate_limited, service_unavailable, internal_error]}
        message: {type: string}
        requestId: {type: string}
    GenericEnvelope:
      type: object
      required: [status, message, body, requestId]
      properties:
        status: {type: string, const: success}
        message: {type: string, const: SUCCESS}
        body: {}
        requestId: {type: string}
...
    ActionPreview:
      type: object
      required: [actionId, actionType, status, summary, expiresAt]
      properties:
        actionId: {type: string, format: uuid}
        actionType: {type: string}
        status: {type: string, enum: [DRAFT, WAITING_CONFIRMATION, CONFIRMED, EXECUTING, COMPLETED, REJECTED, EXPIRED, FAILED]}
        summary: {type: string}
        parameters: {type: object, additionalProperties: true}
        impact: {type: string}
        expiresAt: {type: string, format: date-time}
        confirmationToken: {type: string, writeOnly: true}
    ActionEnvelope:
      allOf:
        - {$ref: '#/components/schemas/GenericEnvelope'}
        - type: object
          properties:
            body: {$ref: '#/components/schemas/ActionPreview'}
security:
  - BearerAuth: []
    AppCode: []